backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.12.5
python-dotenv==1.2.1
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.5.0
python-multipart==0.0.22
httpx==0.28.1
email-validator==2.3.0
alembic==1.13.1
